refactor: align auth code with FSD boundaries
