+++ src/entities/user/index.ts
@@ -1,4 +1,3 @@
 export * from './model/isAdminRole';
 export * from './model/types';
 export * from './model/useGetMyInfo';
-export * from './model/usePostSignIn';
