+++ src/entities/user/model/types.ts
@@ -6,8 +6,12 @@ export interface UserType {
   id: number;
   email: string;
   name: string;
-  studentNumber: string;
+  studentNumber: string | null;
   role: UserRoleType;
 }
 
 export type UserResponseType = ApiResponse<UserType>;
+
+export type AccountRoleType = UserRoleType;
+export type UserInfoType = UserType;
+export type UserInfoResponseType = UserResponseType;
