+++ src/views/home/ui/HomePage.tsx
@@ -4,8 +4,8 @@ import { useEffect } from 'react';
 
 import { toast } from 'sonner';
 
-import { useGetMyInfo, UserInfoResponseType } from '@/entities/auth';
 import { ProjectsListResponseType, useGetProjects } from '@/entities/project';
+import { useGetMyInfo, UserInfoResponseType } from '@/entities/user';
 import { COOKIE_KEYS } from '@/shared/constants';
 import { useHandleErrorQueryToast } from '@/shared/hooks';
 import { cn, getCookie } from '@/shared/utils';
